services/ingestion/requirements.txt
torch --index-url https://download.pytorch.org/whl/cpu
fastapi~=0.103
uvicorn~=0.23
psycopg2-binary~=2.9
requests~=2.31
langchain~=0.1
pgvector~=0.2
langchain-community~=0.1
sentence_transformers
langchain_huggingface
bs4